@@ -1,50 +1,57 @@
 {
-  "name": "client",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~8.2.7",
-    "@angular/common": "~8.2.7",
-    "@angular/compiler": "~8.2.7",
-    "@angular/core": "~8.2.7",
-    "@angular/forms": "~8.2.7",
-    "@angular/platform-browser": "~8.2.7",
-    "@angular/platform-browser-dynamic": "~8.2.7",
-    "@angular/router": "~8.2.7",
-    "bootstrap": "4.1.1",
-    "font-awesome": "^4.7.0",
-    "ngx-bootstrap": "^6.2.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.5",
-    "@angular/cli": "~8.3.5",
-    "@angular/compiler-cli": "~8.2.7",
-    "@angular/language-service": "~8.2.7",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
-  }
+    "name": "client",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "~10.2.4",
+        "@angular/cdk": "^11.2.0",
+        "@angular/common": "~10.2.4",
+        "@angular/compiler": "~10.2.4",
+        "@angular/core": "~10.2.4",
+        "@angular/forms": "~10.2.4",
+        "@angular/platform-browser": "~10.2.4",
+        "@angular/platform-browser-dynamic": "~10.2.4",
+        "@angular/router": "~10.2.4",
+        "@kolkov/ngx-gallery": "^1.2.3",
+        "bootstrap": "4.1.1",
+        "bootswatch": "^4.5.3",
+        "font-awesome": "^4.7.0",
+        "ng2-file-upload": "^1.4.0",
+        "ngx-bootstrap": "^6.2.0",
+        "ngx-spinner": "^10.0.1",
+        "ngx-timeago": "^2.0.0",
+        "ngx-toastr": "^13.2.0",
+        "rxjs": "~6.6.3",
+        "tslib": "^2.0.0",
+        "zone.js": "~0.10.2"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.1002.1",
+        "@angular/cli": "~10.2.1",
+        "@angular/compiler-cli": "~10.2.4",
+        "@angular/language-service": "~10.2.4",
+        "@types/node": "^12.11.1",
+        "@types/jasmine": "~3.3.8",
+        "@types/jasminewd2": "~2.0.3",
+        "codelyzer": "^5.1.2",
+        "jasmine-core": "~3.5.0",
+        "jasmine-spec-reporter": "~5.0.0",
+        "karma": "~5.0.0",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage-istanbul-reporter": "~3.0.2",
+        "karma-jasmine": "~4.0.0",
+        "karma-jasmine-html-reporter": "^1.5.0",
+        "protractor": "~7.0.0",
+        "ts-node": "~7.0.0",
+        "tslint": "~6.1.0",
+        "typescript": "~4.0.5"
+    }
 }
