@@ -1,33 +1,131 @@
 ﻿using API.Data;
+using API.DTOs;
 using API.Entities;
+using API.Extensions;
+using API.Helper;
+using API.Interfaces;
+using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace API.Controllers
 {
-	[ApiController]
-	[Route("api/[controller]")]
-	public class UsersController : ControllerBase
+	[Authorize]
+	public class UsersController : BaseApiController
 	{
-		private readonly DataContext _dataContext;
+		private readonly IUserRepository _userRepository;
+		private readonly IMapper _mapper;
+		private readonly IPhotoSevice _photoSevice;
 
-		public UsersController(DataContext dataContext)
+		public UsersController(IUserRepository userRepository,
+			IMapper mapper,
+			IPhotoSevice photoSevice)
 		{
-			_dataContext = dataContext;
+			_userRepository = userRepository;
+			_mapper = mapper;
+			_photoSevice = photoSevice;
 		}
 		[HttpGet]
-		public async Task<ActionResult<IEnumerable<AppUser>>> GetUsers()
+		public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers([FromQuery]UserParams userParams)
 		{
-			return await _dataContext.AppUsers.ToListAsync();
+			var user = await _userRepository.GetUserByUsernameAsync(User.GetUsername());
+			userParams.CurrentUsername = User.GetUsername();
+
+			if (string.IsNullOrEmpty(userParams.Gender))
+			{
+				userParams.Gender = user.Gender == "male" ? "female" : "male";
+			}
+
+			var users = await _userRepository.GetMembersAsync(userParams);
+
+			Response.AddPaginationHeader(users.CurrentPage, users.PageSize, users.TotalCount, users.TotalPages);
+
+			return Ok(users);
 		}
-		[HttpGet("{id}")]
-		public async Task<ActionResult<AppUser>> GetUsers(int id)
+		[HttpGet("{username}",Name ="GetUser")]
+		public async Task<ActionResult<MemberDto>> GetUsers(string username)
 		{
-			return await _dataContext.AppUsers.FindAsync(id);
+			return await _userRepository.GetMemberUsernameAsync(username);
+		}
+		[HttpPut]
+		public async Task<ActionResult> UpdateUser(MemberUpdateDto memberUpdateDto)
+		{
+			var username = User.GetUsername();
+			var user = await _userRepository.GetUserByUsernameAsync(username);
+
+			_mapper.Map(memberUpdateDto, user);
+			_userRepository.Update(user);
+			if (await _userRepository.SaveAllAsync())
+				return NoContent();
+			return BadRequest("Failed to update User");
+		}
+		[HttpPost("add-photo")]
+		public async Task<ActionResult<PhotoDto>> AddPhoto(IFormFile file)
+		{
+			var username = User.GetUsername();
+			var user = await _userRepository.GetUserByUsernameAsync(username);
+			var result = await _photoSevice.AddPhotoAsync(file);
+			if (result.Error != null )
+				return BadRequest(result.Error.Message);
+
+			var photo = new Photo
+			{
+				Url = result.SecureUrl.AbsoluteUri,
+				PublicId = result.PublicId
+			};
+			if (user.Photos.Count == 0)
+				photo.IsMain = true;
+			user.Photos.Add(photo);
+			if (await _userRepository.SaveAllAsync())
+			{
+				//return CreatedAtRoute("GetUser", _mapper.Map<PhotoDto>(photo));
+				return CreatedAtRoute("GetUser",new {username= user.UserName}, _mapper.Map<PhotoDto>(photo));
+			}	
+				
+
+			return BadRequest("Problem adding photo");
+		}
+		[HttpPut("set-main-photo/{photoId}")]
+		public async Task<ActionResult> SetMainPhoto(int photoId)
+		{
+			var user = await _userRepository.GetUserByUsernameAsync(User.GetUsername());
+			var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
+			if (photo.IsMain)
+				return BadRequest("This is already your main photo");
+			var currentMain = user.Photos.FirstOrDefault(x => x.IsMain);
+			if (currentMain!= null)
+				currentMain.IsMain = false;
+			photo.IsMain = true;
+			if (await _userRepository.SaveAllAsync())
+				return NoContent();
+			return BadRequest("Failed to set main Photo");
+		}
+		[HttpDelete("delete-photo/{photoId}")]
+		public async Task<ActionResult> DeletePhoto(int photoId)
+		{
+			var user = await _userRepository.GetUserByUsernameAsync(User.GetUsername());
+			var photo = user.Photos.FirstOrDefault(x => x.Id == photoId);
+			if (photo == null)
+				return NotFound();
+			if (photo.IsMain)
+				return BadRequest("You cannot delete your main photo");
+			if (!string.IsNullOrEmpty(photo.PublicId))
+			{
+				var result = await _photoSevice.DeletePhotoAsync(photo.PublicId);
+				if (result.Error!= null)
+					return BadRequest(result.Error.Message);
+			}
+			user.Photos.Remove(photo);
+			if (await _userRepository.SaveAllAsync())
+				return Ok();
+			return BadRequest("Cannot delete photo");
 		}
 	}
 }
