@@ -1,4 +1,5 @@
-﻿using System;
+﻿using API.Extensions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -9,6 +10,20 @@ namespace API.Entities
 	{
 		public int Id { get; set; }
 		public string UserName { get; set; }
-
+		public byte[] PasswordHash { get; set; }
+		public byte[] PasswordSalt { get; set; }
+		public DateTime DateOfBirth { get; set; }
+		public string KnownAs { get; set; }
+		public DateTime Created { get; set; } = DateTime.Now;
+		public DateTime LastActive { get; set; } = DateTime.Now;
+		public string Gender { get; set; }
+		public string Introduction { get; set; }
+		public string LookingFor { get; set; }
+		public string Interests { get; set; }
+		public string City { get; set; }
+		public string Country { get; set; }
+		public virtual ICollection<Photo> Photos { get; set; }
+		public virtual ICollection<UserLike> LikedByUsers { get; set; }
+		public virtual ICollection<UserLike> LikedUsers { get; set; }
 	}
 }
