@@ -1,3 +1,5 @@
+import { AccountService } from './_services/account.service';
+import { User } from './_models/user';
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 
@@ -9,18 +11,17 @@ import { Component, OnInit } from '@angular/core';
 export class AppComponent implements OnInit {
 	title = 'Dating app title';
 	users :any;
-	constructor(private http: HttpClient){
+	constructor(private accountService:AccountService){
 		
 	}
 	ngOnInit() {
-		this.getUser();
+		this.setCurrentUser();
 	}
-	getUser(){
-		this.http.get('https://localhost:44356/api/Users/').subscribe(response =>{
-			this.users = response;
-		},error =>{
-			console.log(error);
-		})
+	/// gán user hiện tại = user từ localStorage
+	setCurrentUser()
+	{
+		const user:User = JSON.parse(localStorage.getItem('user'));
+		this.accountService.setCurrentUser(user);
 	}
 }
 
